+++ Sources/Flows/OAuth2ClientCredentials.swift
@@ -36,7 +36,7 @@ open class OAuth2ClientCredentials: OAuth2 {
 	override open func doAuthorize(params inParams: OAuth2StringDict? = nil) {
 		self.obtainAccessToken(params: inParams) { params, error in
 			if let error = error {
-				self.didFail(withError: error)
+				self.didFail(with: error.asOAuth2Error)
 			}
 			else {
 				self.didAuthorize(withParameters: params ?? OAuth2JSON())
@@ -72,7 +72,7 @@ open class OAuth2ClientCredentials: OAuth2 {
 	
 	- parameter callback: The callback to call after the process has finished
 	*/
-	public func obtainAccessToken(params: OAuth2StringDict? = nil, callback: ((_ params: OAuth2JSON?, _ error: Error?) -> Void)) {
+	public func obtainAccessToken(params: OAuth2StringDict? = nil, callback: ((_ params: OAuth2JSON?, _ error: OAuth2Error?) -> Void)) {
 		do {
 			let post = try accessTokenRequest(params: params).asURLRequest(for: self)
 			logger?.debug("OAuth2", msg: "Requesting new access token from \(post.url?.description ?? "nil")")
@@ -85,12 +85,12 @@ open class OAuth2ClientCredentials: OAuth2 {
 					callback(params, nil)
 				}
 				catch let error {
-					callback(nil, error)
+					callback(nil, error.asOAuth2Error)
 				}
 			}
 		}
 		catch let error {
-			callback(nil, error)
+			callback(nil, error.asOAuth2Error)
 		}
 	}
 }